adk_custom_connector/pyproject.toml
[project]
name = "adk_custom_connector"
version = "0.1.0"
description = "A Google ADK Python demo agent for Jira Integration Connectors"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.1.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-gcp-logging>=1.6.0",
    "opentelemetry-exporter-gcp-monitoring>=1.7.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.26.0",
    "opentelemetry-instrumentation>=0.59b0",
    "opentelemetry-instrumentation-google-genai>=0.4b0",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "opentelemetry-sdk>=1.38.0",
]

[tool.uv]
package = false
